API Documentation

LatticeHamiltonians.LatticeHamiltonians — Module

Functions and structures for constructing and applying Hamiltonians on lattice systems.

LatticeHamiltonians.LatticeHamiltonian — Type

This structure is used to store information about the Hamiltonian of the system. It includes information about the real-space basis (A, B), number of orbitals (d), system size (L), parameters for potential and hopping functions (params), real space coordinates of orbitals within a unit cell (r), and a function that applies the Hamiltonian to an input wavefunction (apply!).

LatticeHamiltonians.ham_expr — Method

ham_expr(V, T, dim)

Combine the expressions for the diagonal and hopping terms to generate a block of expressions that applies the full Hamiltonian.

LatticeHamiltonians.loop_periodic — Method

loop_periodic(hops, ex; s = :n)
loop_periodic_diag(dim, d, ex; s = :n)
loop_periodic(s, hop, ex, j)

Generate an Expr block for loops iterating over lattice sites in a periodic system.

LatticeHamiltonians.make_apply — Method

make_apply(params, V, T, dim)

Generate an Expr that defines a function to apply the Hamiltonian given parameters for the potential and hopping terms (V and T).

LatticeHamiltonians.make_diag_expr — Method

make_diag_expr(Vs; s = :n)

Constructs the diagonal part of the Hamiltonian matrix in an efficient way by using metaprogramming features of Julia. It takes the on-site potentials Vs and generates a corresponding Expr block, which, when evaluated, would perform the operations of the diagonal part of the Hamiltonian.

LatticeHamiltonians.make_hop_expr — Method

make_hop_expr(is, js, Vs, dim; s = :n)

Constructs the off-diagonal part of the Hamiltonian matrix by generating an Expr block, which performs the operations of the off-diagonal part of the Hamiltonian.

LatticeHamiltonians.site_expr — Method

site_expr(hops[, j])

Recursively generate an expression for calculating the site site_expr is overloaded for different numbers of arguments.

LatticeHamiltonians.@lattice_hamiltonian — Macro

Define a Hamiltonian using a convenient mini domain specific language (described below). The user can provide an expression that defines the parameters for the potential and hopping terms (params, V, T, L). The macro generates code that constructs a LatticeHamiltonian object and a function to apply the Hamiltonian.