API Documentation

Public API

Internal API

API Documentation

Public API

The are the types and functions the user is expected to use directly.

LatticeHamiltonians.LatticeHamiltonians — Module

Functions and structures for constructing and applying Hamiltonians on lattice systems.

LatticeHamiltonians.LatticeHamiltonian — Type

LatticeHamiltonian{real_dim,lattice_dim,F}

Stores information about the Hamiltonian of the system, so that matrix-vector multiplication can be performed efficiently.

This object should not be initialized by the user. Instead, for construction of the Hamiltonian, see the @lattice_hamiltonian macro.

Fields

A::SMatrix{real_dim,lattice_dim,Float64}: Matrix whose columns are the primitive lattice vectors

B::SMatrix{real_dim,lattice_dim,Float64}: Matrix whose columns are the reciprocal lattice vectors

d::Int: Number of orbitals per site

L::MVector{lattice_dim,Int}: System size

params::Dict{Symbol,ComplexF64}: Parameters for potential and hopping functions

r::Vector{SVector{real_dim,Float64}}: Real-space coordinates of orbitals within a unit cell

apply!::F: Function that applies the Hamiltonian to an input wavefunction

Base.:* — Method

*(H::LatticeHamiltonian, ψ::AbstractVector)

Overloaded matrix multiplication operator for LatticeHamiltonians using the mul! function.

LinearAlgebra.mul! — Method

mul!(ψout::AbstractVector, H::LatticeHamiltonian, ψin::AbstractVector)

Efficiently perform the matrix multiplaction H*ψin, writing the result to ψout.

LatticeHamiltonians.@lattice_hamiltonian — Macro

lattice_hamiltonian(input)

Define a Hamiltonian using a convenient mini domain specific language (described below). The user can provide an expression that defines the parameters for the potential and hopping terms (params, V, T, L). The macro generates code that constructs a LatticeHamiltonian object and a function to apply the Hamiltonian.

Domain Specific Language

lattice_hamiltonian uses a simple domain specific language to define the Hamiltonian. It expects the following types of expressions, supplied in any order.

REQUIRED L = [L1, L2, ...]: A vector of integers, of length lattice_dim that define the size of the system. The length of this vector defines the dimensionality of the system. And each entry is the the number of unit cells in that direction.

REQUIRED V = [V1, V2, ...]: A vector of numbers of length d, that define the on-site potential for each orbital within a unit cell.

Expressions of the orm (δ1, δ2, ...) -> T, with δi integers and T a (d\times d) matrix: The hopping matrix for to hop by (δ1, δ2, ...) unit cells.

Examples

The following creates a Su-Schrieffer-Heeger (SSH) Hamiltonian on a 1D lattice, with 10 unit cells.

H_ssh = @lattice_hamiltonian begin
    L = [10]
    V = [0.0, 0.0]
    (0) -> [0 t1
            t1 0]
    (1) -> [0 t2
            0 0]
    (-1) -> [0 0
            t2 0]
    t1 = 1.0
    t2 = 2.0
end

Internal API

Documentation for internal types and functions are provided here to provide quick reference for developers.

LatticeHamiltonians.LiteralOrSymbolic — Type

LiteralOrSymbolic

Possible types of parameters for the potential and hopping functions.

LatticeHamiltonians.SparseEntry — Type

SparseEntry{T}

A tuple of the form (i, j, val) that describes a non-zero element of a Matrix{T}.

LatticeHamiltonians.bound_nonzero — Method

bound_nonzero(vol, d, T)

Provide an upper bound on the number of non-zero elements in the Hamiltonian matrix, based on the provided hoppings T, the number of orbitals d, and the volume of the system vol.

LatticeHamiltonians.extract_matrix_elements — Method

extract_matrix_elements(hops, exprV, params)

Given a vector of hopping expressions, an expression for the on-site potential, and a dictionary of parameters, returns a tuple of the form (T, V) where T is an array of tuples (rows, cols, values) that describe the non-zero elements of the hopping matrix, and V is a vector of the on-site potentials.

LatticeHamiltonians.ham_expr — Method

ham_expr(V, T, dim)

Combine the expressions for the diagonal and hopping terms to generate a block of expressions that applies the full Hamiltonian.

LatticeHamiltonians.loop_periodic — Method

loop_periodic(hops, ex; s = :n)
loop_periodic_diag(dim, d, ex; s = :n)
loop_periodic(s, hop, ex, j)

Generate an Expr block for loops iterating over lattice sites in a periodic system.

LatticeHamiltonians.make_apply — Method

make_apply(params, V, T, dim)

Generate an Expr that defines a function to apply the Hamiltonian given parameters for the potential and hopping terms (V and T).

LatticeHamiltonians.make_diag_expr — Method

make_diag_expr(Vs; s = :n)

Constructs the diagonal part of the Hamiltonian matrix in an efficient way by using metaprogramming features of Julia. It takes the on-site potentials Vs and generates a corresponding Expr block, which, when evaluated, would perform the operations of the diagonal part of the Hamiltonian.

LatticeHamiltonians.make_hop_expr — Method

make_hop_expr(is, js, Vs, dim; s = :n)

Constructs the off-diagonal part of the Hamiltonian matrix by generating an Expr block, which performs the operations of the off-diagonal part of the Hamiltonian.

LatticeHamiltonians.nonzero_elements — Method

nonzero_elements(matrix)

Takes an expression describing a matrix and returns a named tuple of vectors (rows, cols, vals) where each vector respectively contains the row, column, and value of a nonzero element in the matrix. All numbers are converted to ComplexF64 numbers.

LatticeHamiltonians.parse_hopping — Method

parse_hopping(hop::Expr, params::Dict{Symbol,ComplexF64})

Extract the key and value from a hopping expression of the form

(δ1, δ2, ...) -> [t1 t2 ...; t3 t4 ...; ...]

and return a pair of the form

key => (rows, cols, values)

where key is a vector of integers that describe the hopping vector, and rows, cols, and values are vectors that describe the non-zero elements of the hopping matrix.

The hopping matrix itself may be a singleton, a matrix literal, or a tuple of vectors (rows, cols, values).

LatticeHamiltonians.site_expr — Method

site_expr(hops[, j])

Recursively generate an expression for calculating the site site_expr is overloaded for different numbers of arguments.

LatticeHamiltonians.symbols_to_lookups — Method

symbols_to_lookups(expr, params::Dict{Symbol,ComplexF64})

Attempts to convert terms to a canonical form:

All literal numbers are converted to ComplexF64

All symbols that are keys in params are converted to lookups, i.e. t1 becomes params[:t1].

Everything else is left as is.