API Documentation

Public API

Internal API

API Documentation

Public API

These are the types and functions the user is expected to use directly.

LatticeHamiltonians.LatticeHamiltonians — Module

Functions and structures for constructing and applying Hamiltonians on lattice systems.

LatticeHamiltonians.LatticeHamiltonian — Type

LatticeHamiltonian{real_dim,lattice_dim,F,S}

Stores information about the Hamiltonian of the system, so that matrix-vector multiplication can be performed efficiently.

This object should not be initialized by the user. Instead, for construction of the Hamiltonian, see the @lattice_hamiltonian macro.

Fields

A::SMatrix{real_dim,lattice_dim,Float64}: Matrix whose columns are the primitive lattice vectors

B::SMatrix{real_dim,lattice_dim,Float64}: Matrix whose columns are the reciprocal lattice vectors

d::Int: Number of orbitals per site

L::MVector{lattice_dim,Int}: System size

params::Dict{Symbol,ComplexF64}: Parameters for potential and hopping functions

r::Vector{SVector{real_dim,Float64}}: Real-space coordinates of orbitals within a unit cell

apply!::F: Function that applies the Hamiltonian to an input wavefunction

sparse::S: Function that constructs the sparse matrix representation of the Hamiltonian

LatticeHamiltonians.@lattice_hamiltonian — Macro

lattice_hamiltonian(input)

Define a Hamiltonian using a convenient mini domain specific language (described below). The user can provide an expression that defines the parameters for the potential and hopping terms (params, V, T, L). The macro generates code that constructs a LatticeHamiltonian object and a function to apply the Hamiltonian.

Domain Specific Language

lattice_hamiltonian uses a simple domain specific language to define the Hamiltonian. It expects the following types of expressions, supplied in any order.

REQUIRED L = [L1, L2, ...]: A vector of integers, of length lattice_dim that define the size of the system. The length of this vector defines the dimensionality of the system. And each entry is the the number of unit cells in that direction.

REQUIRED V = [V1, V2, ...]: A vector of numbers of length d, that define the on-site potential for each orbital within a unit cell.

Expressions of the orm (δ1, δ2, ...) -> T, with δi integers and T a (d\times d) matrix: The hopping matrix for to hop by (δ1, δ2, ...) unit cells.

Examples

The following creates a Su-Schrieffer-Heeger (SSH) Hamiltonian on a 1D lattice, with 10 unit cells.

H_ssh = @lattice_hamiltonian begin
    L = [10]
    V = [0.0, 0.0]
    (0) -> [0 t1
            t1 0]
    (1) -> [0 t2
            0 0]
    (-1) -> [0 0
            t2 0]
    t1 = 1.0
    t2 = 2.0
end

Internal API

Documentation for internal types and functions are provided here to provide quick reference for developers.

LatticeHamiltonians.LiteralOrSymbolic — Type

LiteralOrSymbolic

Possible types of parameters for the potential and hopping functions.

LatticeHamiltonians.HamiltonianBuilder — Type

HamiltonianBuilder{real_dim,lattice_dim}

Contains the information needed to build a Hamiltonian for a lattice system.

Fields

The fields here are subset of those on LatticeHamiltonian and are used to construct the Hamiltonian.

See also: @builder, build

LatticeHamiltonians.LatticeVector — Type

LatticeVector{lattice_dim}

A vector of integers that describes a lattice vector in a lattice_dim dimensional lattice.

LatticeHamiltonians.SparseEntry — Type

SparseEntry{T}

A tuple of the form (i, j, val) that describes a non-zero element of a Matrix{T}.

Base.:* — Method

*(H::LatticeHamiltonian, ψ::AbstractVector)

Overloaded matrix multiplication operator for LatticeHamiltonians using the mul! function.

LatticeHamiltonians.bound_nonzero — Method

bound_nonzero(vol, d, T)

Provide an upper bound on the number of non-zero elements in the Hamiltonian matrix, based on the provided hoppings T, the number of orbitals d, and the volume of the system vol.

LatticeHamiltonians.build — Method

build(builder::HamiltonianBuilder{real_dim,lattice_dim})

Create a LatticeHamiltonian from a HamiltonianBuilder.

LatticeHamiltonians.dim_span_var — Method

dim_span_var(j)

Generate a symbol for the span of the lattice dimension in the generated code. This describes how much the linear index changes when traversing all sites in this dimension. For example, in a 2D lattice, we might generate Lprod1, Lprod2 for the span of the lattice dimensions. Suppose that there are d=3 orbitals on-site, and the lattice dimensions are L1=2 and L2=3. Then the linear index changes by d * L1 when traversing all sites in the first dimension, and by d * L1 * L2 when traversing all sites in the second dimension, so the spans are 6 and 18 respectively.

LatticeHamiltonians.extract_potential — Method

extract_potential(exprV, dim, d, params::Dict{Symbol,ComplexF64})

Given an onsite hopping seperate the potential from the site-local hopping matrix. Allowed forms for the right hand side are the same as parse_hopping.

Besides the expression, also takes as input the lattice dimension dim, the number of orbitals per site d, and the parameter dictionary params.

Return as tuple of the form (V, hop).

The potential is returned as a vector of LiteralOrSymbolic values. If there is no non-zero onsite hopping, hop=nothing, otherwise hop is a Tuple of the form

(rows, cols, values)

extract_potential(:([Δ t; t -Δ], 1, 2, Dict{Symbol,ComplexF64}(:Δ => 1.0))

LatticeHamiltonians.foreach_element — Method

foreach_element(f :: Function, expr::Expr)

Generic iteration over elements of a matrix or vector expression. f should be a function of the form f(ri, ci, elem) where ri and ci are the row and column indices of the element, and elem is the value.

LatticeHamiltonians.ham_expr — Method

ham_expr(V, T, dim; sparse=false)

Combine the expressions for the diagonal and hopping terms to generate a block of expressions that applies the full Hamiltonian.

The behavior of the function can be controlled by the sparse keyword argument:

when sparse is false (default), the function explicitly performs a matrix-vector multiplication on a vector ψin (which is assumed to be defined in the enclosing scope) and stores the results in ψout.

when sparse is true, the function generates a sparse matrix representation of the Hamiltonian which can be used to instantiate a sparse matrix.

When implementing matrix-vector multiplication directly, the incoming and outgoing vectors are flattened into 1D arrays. This effectively corresponds to a Kroneker product of the lattice dimensions. Because Julia uses column-major ordering, the indices of the multi-dimensional array are ordered fastest to slowest changing: [o, i1, i2, ..., in] where o is the oribital index and i1, i2, ..., in are the lattice indices in dimensions 1, 2, ..., n.

LatticeHamiltonians.isliteral — Method

isliteral(expr)

Check if an expression is a literal number or a literal complex number, using a simple heuristic. In general, any expression that is a number or a simple arithmetic operation on numbers is considered literal.

Note

Currently only supports numbers, im, and the arithmetic operations +, -, *, /, and ^. Function calls like sqrt(2) are not considered literal.

LatticeHamiltonians.isonsite — Method

isonsite(expr::Expr)

Given an expression of the form δ -> [t1 t2 ...; t3 t4 ...; ...], checks if the hopping is on-site, i.e. if δ is zero.

On site values for δ are

0 -> ...
(0) -> ...
(0, 0, ...) -> ...

LatticeHamiltonians.loop_site_pairs — Method

loop_site_pairs(hops, ex, axis)

The main method of loop_site_pairs works by recursively calling in to this method. Each call to this method generates a block of for loops for a single lattice dimension, and then calls the next dimension recursively, decreasing axis by 1. When we have reached the last dimension, we insert the expression ex at each site.

LatticeHamiltonians.loop_site_pairs — Method

loop_site_pairs(hops, ex)

Recursively generate an Expr block which loops over all pairs of lattice sites separated by the lattice vector hops and insert the expression ex for each pair.

julia> using LatticeHamiltonians: loop_periodic

julia> using MacroTools: prettify

julia> prettify(loop_periodic([1 0 1], :(expr)))

quote
    i_out = 0
    i_in = d * (1 + (0 + 1 * L[2]) * L[1])
    for n3 = 1:L[3] - 1
        for n2 = 1:L[2]
            for n1 = 1:L[1] - 1
                expr
            end
            i_in -= Lprod1
            for n1 = L[1] - 0:L[1]
                expr
            end
            i_in += Lprod1
        end
    end
    i_in -= Lprod3
    for n3 = L[3] - 0:L[3]
        for n2 = 1:L[2]
            for n1 = 1:L[1] - 1
                expr
            end
            i_in -= Lprod1
            for n1 = L[1] - 0:L[1]
                expr
            end
            i_in += Lprod1
        end
    end
    i_in += Lprod3
end

LatticeHamiltonians.loop_sites — Method

loop_sites(dim, ex)

Recursively generate an Expr block which loops over lattice sites and inserts the expressions ex at each site.

LatticeHamiltonians.make_apply — Method

make_apply(V, T, dim)

Generate an Expr that defines a function to apply the Hamiltonian given parameters for the potential and hopping terms (V and T).

LatticeHamiltonians.make_diag_expr — Method

make_diag_expr(Vs; s = :n)

Constructs the diagonal part of the Hamiltonian matrix in an efficient way by using metaprogramming features of Julia. It takes the on-site potentials Vs and generates a corresponding Expr block, which, when evaluated, would perform the operations of the diagonal part of the Hamiltonian.

If sparse is set to true, the function generates an Expr block that constructs the sparse matrix representation of the diagonal part of the Hamiltonian.

LatticeHamiltonians.make_hop_expr — Method

make_hop_expr(is, js, hs, dim; s = :n, sparse=false)

Constructs the off-diagonal part of the Hamiltonian matrix by generating an Expr block, which performs the operations of the off-diagonal part of the Hamiltonian.

If sparse is set to true, the function generates an Expr block that constructs the sparse matrix representation of the off-diagonal part of the Hamiltonian.

LatticeHamiltonians.make_sparse — Method

make_sparse(V, T, dim)

Generate an Expr that defines a function to construct the sparse matrix representation of the Hamiltonian given parameters for the potential and hopping terms (V and T).

LatticeHamiltonians.matrix_element — Method

matrix_element(h::Function; site)
matrix_element(h::ComplexF64; site)
matrix_element(h; site)

Generate an expression for the orbital matrix element h at a given lattice site. If h is a function, it is evaluated at the given site, i.e., h(n1, n2, n3...). If h is a number, it is returned as a constant. If h is a symbol, it is looked up from the params dictionary.

LatticeHamiltonians.nonzero_elements — Method

nonzero_elements(matrix)

Takes an expression describing a matrix and returns a named tuple of vectors (rows, cols, vals) where each vector respectively contains the row, column, and value of a nonzero element in the matrix. All numbers are converted to ComplexF64 numbers.

If passed a vector, returns the nonzero elements of the corresponding diagonal matrix.

LatticeHamiltonians.orbital_dim — Method

orbital_dim(T)

compute the number of orbitals as the maximum index of the orbital hopping matrices

LatticeHamiltonians.parse_hopping — Method

parse_hopping(hop::Expr, params::Dict{Symbol,ComplexF64})

Extract the key and value from a hopping expression of the form

(δ1, δ2, ...) -> [t1 t2 ...; t3 t4 ...; ...]

and return a pair of the form

key => (rows, cols, values)

where key is a vector of integers that describe the hopping vector, and rows, cols, and values are vectors that describe the non-zero elements of the orbital hopping matrix.

The hopping matrix itself may be a singleton, a matrix literal, or a tuple of vectors (rows, cols, values).

LatticeHamiltonians.site_expr — Method

site_expr(hops[, j])

Recursively generate an expression for calculating the site site_expr is overloaded for different numbers of arguments.

LatticeHamiltonians.site_loop_var — Method

site_loop_var(j)

Generate a symbol for the loop variable in the generated code. j indexes the lattice dimension. For example, in a 2D lattice, we might generate n1, n2 for the loop variables.

See also: site_vector_var

LatticeHamiltonians.site_var_vector — Method

site_vector_var(dim)

Generate a vector of symbols for the lattice site in the generated code. dim is the total lattice dimension.

For example, for a 2D lattice,

julia> using LatticeHamiltonians: site_var_vector

julia> site_var_vector(2)
2-element Vector{Symbol}:
 :n1
 :n2

LatticeHamiltonians.symbols_to_lookups — Method

symbols_to_lookups(expr, params::Dict{Symbol,ComplexF64})

Attempts to convert terms to a canonical form:

All literal numbers are converted to ComplexF64

All symbols that are keys in params are converted to lookups, i.e. t1 becomes params[:t1].

Everything else is left as is.

LinearAlgebra.mul! — Method

mul!(ψout::AbstractVector, H::LatticeHamiltonian, ψin::AbstractVector)

Efficiently perform the matrix multiplaction H*ψin, writing the result to ψout.

SparseArrays.sparse — Method

sparse(H::LatticeHamiltonian)

Return the sparse representation of the Hamiltonian matrix.

LatticeHamiltonians.@builder — Macro

@builder

Construct as HamiltonianBuilder using the same DSL as @lattice_hamiltonian.